fix(api): money with kopecks — float ₴ / decimal columns end-to-end
Silpo returns prices like 69.90; we were rounding to int (70), which broke
per-item prices and made cart totals drift from the real Silpo cart.

- SilpoProduct/Candidate prices float
- optimizer totals rounded to 2 decimals
- cart_items + meal_plans money columns → decimal(10,2) (migration) + float casts
- tests updated to float expectations

+++ api/app/Services/Budget/Candidate.php
@@ -12,11 +12,11 @@ final readonly class Candidate
         public string $ingredient,   // нормалізована назва інгредієнта
         public string $sku,          // silpo product id
         public string $title,        // назва товару
-        public int $price,           // ціна за одиницю (акційна/поточна)
+        public float $price,         // ціна за одиницю в ₴ з копійками (акційна/поточна)
         public bool $isPromo = false,
         public bool $isPrivateLabel = false,
         public float $confidence = 1.0, // 0..1 впевненість матчингу
-        public ?int $oldPrice = null,   // ціна до знижки (для розрахунку економії)
+        public ?float $oldPrice = null, // ціна до знижки в ₴ (для розрахунку економії)
         public ?float $packSize = null, // розмір фасовки (у packUnit)
         public ?string $packUnit = null, // 'g' | 'ml' | 'pcs'
     ) {}
